fastapi-reverse-proxy/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
redis==5.0.1
httpx==0.25.2
prometheus-client==0.19.0
PyJWT==2.8.0
email-validator==2.1.0